requirements.txt
Dependencies to install:

pillow
